Big Rework
- Removed routing folder
- Got admin to run on app.py
- Basically retconning admin dashboard entirely to make it more readable
