--- code/admin/__init__.py
@@ -1,1 +1,0 @@
-from .admin import admin_bp
