+++ code/route.py
@@ -1,0 +1,2 @@
+from .admin import admin_bp
+from .app import app_bp
