+++ code/templates/admin/admin-dashboard.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/code/static/css/admin-dashboard.css">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/admin-dashboard.css') }}">
     <title>Admin Dashboard</title>
     <style>
         body {
@@ -17,7 +17,7 @@
 </head>
 
 <body>
-    <iframe src="admin-nav-bar.html" style="width:100%; height: 100%; border:none;"></iframe>
+    <!-- <iframe src="/admin-nav-bar" style="width:100%; height: 100%; border:none;"></iframe> -->
 
     <!-- Display dataset -->
     <!--<iframe id="dataset-frame" src="/../Infinite-Loopers/code/front-end/nav-bar/admin-dashboard/dataset.html" style="width:100%; height:500px; border:none;"></iframe>-->
@@ -30,9 +30,7 @@
             <div class="login-menu">
                 <button id="button">Alert User</button>
                 <h1>Edit User</h1>
-                <input type="text" id="edit-old-username" name="old-username" placeholder="Old Username/Email"><br>
                 <input type="text" id="edit-new-username" name="new-username" placeholder="New Username/Email"><br>
-                <input type="text" id="edit-old-password" name="old-password" placeholder="Old Password"><br>
                 <input type="text" id="edit-new-password" name="new-password" placeholder="New Password"><br>
                 <input type="submit" value="Save" onclick="editUser()">
             </div>
